BC-1833 - moved shared functionality to RoomWrapper component
